I think this checkin eliminates around 15 of Dave's 100 or warning messages. I'm hoping I didn't introduce some new warnings -- which is likely. I will keep chipping away at them since most of them are in code that I wrote:(
git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@934 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: ea6ed6391a9548af7c9396ba36e405fbdc9a6cca

+++ bond_volume/quick_grid/calculators.cxx
@@ -240,7 +240,7 @@ void compute_bond_volume
 	const int *neighPtr = localNeighborList;
 	const double *xOwned = xOverlap;
 	for(size_t p=0;p<num_owned_points;p++, xOwned +=3){
-		size_t numNeigh = *neighPtr; neighPtr++;
+		int numNeigh = *neighPtr; neighPtr++;
 
 		const double *P = xOwned;
 
@@ -268,7 +268,7 @@ void compute_bond_volume
 
 	const int *neighPtr = localNeighborList;
 	const double *P = X;
-	size_t numNeigh = *neighPtr; neighPtr++;
+	int numNeigh = *neighPtr; neighPtr++;
 	for(int n=0;n<numNeigh;n++,neighPtr++,bond_volumes++){
 		int localId = *neighPtr;
 		const double *Q = &xOverlap[3*localId];
